26/11/23

+++ src/Qustions/questionsService.ts
@@ -1,5 +1,5 @@
 import { QuestionsRepository } from './questionsRepository';
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import {
   CreateQuestionDTO,
   QuestionsPaginationDTO,
@@ -9,6 +9,7 @@ import { QuestionEntity } from './Entitys/QuestionEntity';
 import { randomUUID } from 'crypto';
 import { mapQuestions } from './mapQuestions';
 import { helper } from '../helper';
+import { isUUID, IsUUID } from 'class-validator';
 @Injectable()
 export class QuestionsService {
   constructor(private readonly questionsRepository: QuestionsRepository) {}
@@ -28,6 +29,7 @@ export class QuestionsService {
     const questionViewModel = mapQuestions.mapQuestionViewModel(question);
     return questionViewModel;
   }
+
   async getQuestions(questionPaginationDTO: QuestionsPaginationDTO) {
     const pagination = helper.getQuestionPaginationDTO(questionPaginationDTO);
     console.log(pagination);
@@ -35,6 +37,68 @@ export class QuestionsService {
   }
 
   async deleteQuestions(questionId: string): Promise<boolean> {
+    if (isUUID(questionId) === false) {
+      throw new NotFoundException(
+        'questionId not found question /questionService/deleteQuestion',
+      );
+    }
+    const question = await this.questionsRepository.getQuestionId(questionId);
+    if (question.length < 1) {
+      throw new NotFoundException(
+        'questionId not found question /questionService/deleteQuestion',
+      );
+    }
     return this.questionsRepository.deleteQuestion(questionId);
+  }
+
+  async updateQuestion(
+    questionId: string,
+    updateQuestionDTO: CreateQuestionDTO,
+  ) {
+    if (isUUID(questionId) === false) {
+      throw new NotFoundException(
+        'questionId not found question /questionService/updateQuestion',
+      );
+    }
+    const question = await this.questionsRepository.getQuestionId(questionId);
+    if (question.length < 1) {
+      throw new NotFoundException(
+        'questionId not found question /questionService/updateQuestion',
+      );
+    }
+    const updateQuestion = this.questionsRepository.updateQuestion(
+      questionId,
+      updateQuestionDTO,
+    );
+    if (!updateQuestion) {
+      throw new NotFoundException(
+        'question ne obnovilsya, questionService ,update',
+      );
+    }
+    return true;
+  }
+
+  async updatePublishQuestion(questionId: string, published: boolean) {
+    if (isUUID(questionId) === false) {
+      throw new NotFoundException(
+        'questionId not found question /questionService/updateQuestion',
+      );
+    }
+    const question = await this.questionsRepository.getQuestionId(questionId);
+    if (question.length < 1) {
+      throw new NotFoundException(
+        'questionId not found question /questionService/updateQuestion',
+      );
+    }
+    const updateQuestion = this.questionsRepository.updatePublishQuestion(
+      questionId,
+      published,
+    );
+    if (!updateQuestion) {
+      throw new NotFoundException(
+        'question ne obnovilsya, questionService ,update',
+      );
+    }
+    return true;
   }
 }
